+++ CHANGELOG.md
@@ -7,6 +7,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 ## [Unreleased] - TBD
 ### Added
 - Added `gcpy/benchmark/modules/benchmark_gchp_stats.py`
+- Added routine to generate summary table to `gcpy/benchmark/modules/benchmark_species_changes.py`
 
 ### Changed
 - Bumped pytest to 9.0.3 in `_setup.py` and environment files; Updated documentation accordingly
@@ -15,7 +16,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - Fixed tuple unpack error in `gcpy/benchmark/modules/benchmark_species_changes.py`
 
 ### Removed
-- Removed the Advected column from the wiki tables in `gcpy/benchmark/modules/benchmark_species_changes.py` since `Is_Advected` is no longer a species_database field
+- Removed the Advected column from the wiki tables in `gcpy/benchmark/modules/benchmark_species_changes.py`
 
 ## [1.8.0] - 2026-08-10
 ### Added